fix(errors): Custom RxJS errors now all have a call stack (#5686)
* fix(errors): Custom RxJS errors now all have a call stack

NOTE: Because we have to workaround compilation-to-ES5 inadequacies until call stack locations will show as being inside of the constructor implementation at the top level. This will go away as we move the community to ES2015+

BREAKING CHANGE: Tests that are written with naive expectations against errors may fail now that errors have a proper `stack` property. In some testing frameworks, a deep equality check on two error instances will check the values in `stack`, which could be different.

fixes #4250

* refactor: Improve error class creator

- Adds tests

* chore: update side-effects golden files

* chore: address comments

* chore: update side-effects golden files again

* chore: I hate ts-api-guardian sometimes