refactor(toArray): reduce complexity a little.
There was a situation where the `toArray` implementation was relying on a quirk of `reduce` and a trick with the `index` in order to ensure that two subscriptions to the same `toArray`-ed observable would not mutate the same underlying array.

This change gets around that problem in a more straight-forward manner, if at the cost of a few bytes.