fix(zip): zip now accepts an array of arguments like its counterparts
`forkJoin` and `combineLatest` both allow an array of arguments to be passed to it like so: `forkJoin([a$, b$, c$])`. Now `zip` does the same: `zip([a$, b$, c$])`

BREAKING CHANGE: Zipping a single array will now have a different result. This is an extreme corner-case, because it is very unlikely that anyone would want to zip an array with nothing at all.