refactor(expand): use mergeInternals
- Adds a moderate amount of complexity to `mergeInternals` to support `expand`.
- `expand` now based off of the same internals as `mergeMap` and `mergeScan`.
- `deprecates` the scheduler argument to `expand` with an explanation.

NOTE: While I'm not fond of the number of arguments to `mergeInternals`, it's an _internal_ API, and given that it's not public, I'd like to refrain from creating named parameters via a configuration object just to make our internal code more readable, especially when this is relatively hot code. Instead, I've opted to make sure everything is well commented.