fix(iif): No longer allow accidental undefined arguments
- Updates types
- Ensures that result arguments must be passed
- Ensures narrower type when conditional result is known to be `true`, or known to be `false`.

BREAKING CHANGE: `iif` will no longer allow result arguments that are `undefined`. This was a bad call pattern that was likely an error in most cases. If for some reason you are relying on this behavior, simply substitute `EMPTY` in place of the `undefined` argument. This ensures that the behavior was intentional and desired, rather than the result of an accidental `undefined` argument.