fix(window): final window stays open until source complete
Resolves an issue where the windowBoundary complete would complete the resulting observable.

BREAKING CHANGE: The windowBoundaries observable no longer completes the result. It was only ever meant to notify of the window boundary. To get the same behavior as the old behavior, you would need to add an `endWith` and a `skipLast(1)` like so:  `source$.pipe(window(notifier$.pipe(endWith(true))), skipLast(1))`.