docs: Document differences between 6.x and 7.x golden files
Adds a document outlining all discovered differences between version 6 and version 7 as they can be found while diffing the api_guardian snapshots. This document should be an exhaustive list of changes and we should modify it ongoing if we think of anything else that has changed that cannot be found here yet.