feat: add config object to connectable (#6267)
* feat: add config object to connectable

* chore: update api_guardian

* chore: typo

BREAKING CHANGE: Our very new api, `connectable`, now takes a configuration object instead of just the `Subject` instance. This was necessary to make sure it covered all use cases for what we were trying to replace in the deprecated multicasting operators. Apologies for the late-in-the-game change, but we know it's not widely used yet (it's new in v7), and we want to get it right.

Co-authored-by: Ben Lesh <ben@benlesh.com>