fix(takeWhile): reentrant errors and completions behave properly
- Resolves an issue where a reentrant error notification would short circuit the completion.
- Adds additional tests.

Related #5487

BREAKING CHANGE: If a the source synchronously errors after it recieves a completion notification, the error will no longer be emitted. This is a bug fix, but may be a breaking change for those relying on this behavior. If you need to mimic the behavior, you'll need to throw the error before the takeWhile notifier is notified.