fix: use "unknown" type for operators that ignore input values
These operators ignore the values produced by the source Observable, so
they don't need the dynamic typing that "any" would provide for
accessing the values.  Using "unknown" instead of "any" avoids warnings
from eslint rules like @typescript-eslint/no-unsafe-argument in
application code that uses the operators, and it's also a clearer way to
express that the operators don't know or care what the source type is.