fix(ajax): crossDomain flag deprecated and properly reported to consumers (#6710)
* fix(ajax): crossDomain flag deprecated and properly reported to consumers

Fixes an issue where the `crosDomain` flag was being incorrectly reported to users via error objects and response objects as defaulting to `true`, when it was in fact defaulting to `false`.

Deprecates the `crossDomain` flag in favor of allowing the configuration of the request and the browser to dictate whether or not a preflight request is necessary. Adds deprecation messages with advice about how to force CORS preflights. Ultimately, the boolean flag does not make sense, as there are a lot of factors that dictate CORS preflight use and they may vary by browser/environment.

Resolves #6663

* chore: Update side effects file

* chore: update side effects files