feat: Add allValuesFrom function
This is similar to firstValueFrom and lastValueFrom, but the promise
resolves to an array of all the values emitted by the observable,
instead of just the first or last.  The implementation is trivial and
applications can easily get the same effect by directly using toArray
with firstValueFrom/lastValueFrom, but a named function makes it more
clear to a reader what the combination does.