perf: removed code that would `bind` functions passed with observers to `subscribe`. (#6815)
* refactor(SafeSubscriber): optimize perf for ordinary observers

No longer require function binding if we aren't using the deprecated next context. This should improve performance in the common path of consumers subscribing with an object or even with a function.

Adds a simple class `ConsumerObserver` which is mostly meant to optimize the number of function refrences created. We should never expose this externally.

Related #6783

* chore: update comments

* refactor(Subscriber): reduce property access