feat(shareReplay): use another observable to control resets
This feature, first introduced for the `share` operator, is equally as desirable for the `shareReplay` operator. It allows one to, for example, set up a delayed reset. For Observables that are a little expensive to set up, but not so expensive that it makes sense to keep them around forever.