feat(onErrorResumeNextWith): renamed `onErrorResumeNext` and exported from the top level (#6755)
* feat(onErrorResumeNextWith): renamed `onErrorResumeNext` and exported from top level

The `onErrorResumeNext` operator is deprecated due to a rename. The rename was done so we can move all operator exports to the top level `rxjs` export site. We were not currently exporting that operator.

This commit also refactors to logic to be based on the creation function, which is more widely used

* docs: Update import information for onErrorResumeNext, fix info for partition