feat: RxJS updated to TypeScript 4.9
* chore(build): Update TypeScript to 4.9.4

* chore: Mocha test output updated to ES2022.

There was an issue where it was using some sort of poorly transpiled version of async generators that didn't finalize in a deterministic way, causing some tests to fail.

* chore: Fix ReadableStreamLike for TS 4.9

BREAKING CHANGE: RxJS is now being built with TypeScript 4.9 and we assume that TS 4.9 or higher will be used with our types. At the time of this writing the risk is around the existence of the `ReadableStream` type in TypeScript. If that's not supported in your version of TypeScript, you might be able to work around this issue by defining a global ReadableStream type.