chore: update ts-node, mocha, chai, and sinon-chai
We had to update ts-node for the newer version of TypeScript. This PR showed an issue where when we used `??=` in the code, CI would fail for Node 14... but it would fail with a nonsense error. Trial and elimination showed that the error was caused when the syntax error was being handled by mocha in ts-node in Node 14. It caused an error like `Unknown pluginType "ui"`. Updating all of these dependencies resolved the issue so I could see the actual error.