fix(buffer): subscribe to the closing notifier before the source (#7185)
BREAKING CHANGE: Weird patterns where an observable is used to notify buffers of itself will result in a first emission of `[]`. If you need to buffer an array with itself, just use `map(x => [x])`. If that doesn't work, feel free to file a use case as an issue. `skip(1)` would be the other workaround.