perf(Subject): subscription, unsubscription, and nexting improved ~4x (#7240)
* fix(subject): Faster subscriptions management in Subject

The current implementation of the Subject was backing itself on an array of observers. When one observer unsubscribed itself, it was performing a linear scan on the array of observers and dropping the item from the array of observers. The algorithm complexity of this operation was O(n) with n being the number of observers connected to the Subject.

It caused code like:

```js
const allSubscriptions = [];
const source = new Subject();
for (let index = 0 ; index !== 1_000_000 ; ++index) {
  allSubscriptions.push(source.subscribe()); // rather quick
}
for (const subscription of allSubscriptions) {
  subscription.unsubscribe(); // taking very long
}
```

The proposed approach consists into changing our backing collection (an array) into a Map. But we lose somehow the set capability on subject.observers (might possibly be patched in some way).

Following that change the command `cross-env TS_NODE_PROJECT=tsconfig.mocha.json mocha --config spec/support/.mocharc.js "spec/**/Subje*-spec.ts"` passed to 452ms from 10s.

* refactor(Subject): some perf optimizations

+ Loops over indices of the observers array rather than creating a new iterator and using for-of for each emission
+ Removes unnecessary dirty boolean in favor of clearing the snapshot, which frees up memory a little
+ Centralizes observers list teardown and ensures it's called during `unsubscribe()`.

---------

Co-authored-by: Ben Lesh <ben@benlesh.com>