feat(repeat): add lastValue as second param when delay is a function
In some cases, we need to use the repeat operator in RxJS to continuously retrieve values such as
tokens. However, sometimes the retrieval interval needs to be determined based on the expiration
time in the last retrieved value. With this change, we have added support for passing the last
retrieved value as the second argument to the delay option when it is a function. This makes it
easier to handle such scenarios.