refactor(bufferWhen): use new `operate` function
Note that this removes setting `destination` to `null`. We're doing this because code will want to unsubscribe from upstream, and _then_ send an error to downstream. This is per operator development guidelines.

Additional context: Since we are closing over `destination` in the `OperatorSubscriber` overrides, nulling out the `destination` in `Subscriber` literally never made a difference in memory, as the `destination` was trapped in memory by closer anyhow. Therefor the change to remove `this.destination = null!;` is no change at all.