feat(rx): A new method of piping has been added (#7229)
* feat(r): A new method of piping has been added

+ Adds `r` a new method of piping observables. This allows any valid observable input to be passed as the first argument, and
 it will implicitly convert the argument to an Observable and pass it to the function that may be present in the second argument, the
 return value of THAT function is then passed an argument to the next function, and so on
+ Corrects the types of `Observable#pipe` such that anything can be returned at any step of the functional chain, however the first pipeable function must accept an `Observable<T>`.
+ Adds dtslint and runtime tests for `r` and `pipe`.

NOTE: This does NOT deprecate`Observable#pipe`. That will be done in a follow up, and we need to define what timeline we'll take to remove that, as it's an API that is broadly used - could be v9, could be v10, could be never. At this point, this is alpha/beta functionality.

BREAKING CHANGE: `Observable#pipe` now allows any pipeable unary function as an argument, just so long as the types properly compose, this means in some cases it will now return `unknown` instead of `Observable<unknown>` the workaround is just to cast the result for those cases. (or you can break your operator piping into smaller pipe sets)

Related #7203

* refactor: rename to `rx`

* docs: fix `r` to `rx` reference.

Co-authored-by: Mladen Jakovljević <jakovljevic.mladen@gmail.com>

* chore: fix description

Co-authored-by: Nicholas Jamieson <nicholas@cartant.com>

* chore: fix return type of part of signature.

Co-authored-by: Mladen Jakovljević <jakovljevic.mladen@gmail.com>

* chore: correct docs to say `rx` and not `r`.

Co-authored-by: Mladen Jakovljević <jakovljevic.mladen@gmail.com>

* chore: address comments

+ Adds some more tests around `Observable#pipe`, ensuring arbitrary unary functions work
+ Adds more dtslint tests for `rx`.
+ Fixes some comments
+ Improves types and runtime for args in `rx` function

---------

Co-authored-by: Mladen Jakovljević <jakovljevic.mladen@gmail.com>
Co-authored-by: Nicholas Jamieson <nicholas@cartant.com>