chore: migrate to expecttype eslint rules
This migrates to using an eslint plugin for `$ExpectType` and friends so
we can do it as part of the lint script.

For now, `typescript` had to be added to the root project since the rxjs
site is still on an older typescript which confuses eslint (since it
ends up being hoisted by yarn).