feat(operators): add toModel operator for mapping emitted values to model instances
This change introduces a new `toModel` operator that allows emitted values from
an observable to be mapped to instances of a specified class. The operator takes
a constructor as an argument and transforms the observable stream by mapping the
data to new instances of the given model class.

The feature enables developers to easily map API responses or other emitted
values into strongly-typed objects within an RxJS observable chain.